Addressing review comments for backwards compatibility.
